Stout: Removed warning in Windows implementation of `stat.hpp`.
Review: https://reviews.apache.org/r/47391

+++ 3rdparty/stout/include/stout/os/windows/stat.hpp
@@ -126,7 +126,7 @@ inline Try<long> mtime(const std::string& path)
     return ErrnoError("Error invoking stat for '" + path + "'");
   }
 
-  return s.st_mtime;
+  return static_cast<long>(s.st_mtime);
 }
